+++ .gitignore
@@ -1,0 +1,10 @@
+CMakeCache.txt
+CMakeFiles
+CMakeScripts
+Makefile
+cmake_install.cmake
+install_manifest.txt
+*~
+*#
+cpsumoncli
+libcpsumon/libcpsumon.so
