--- compile.sh
@@ -1,2 +1,0 @@
-#!/bin/sh
-gcc cpsumon.c -gstabs+ -lm -o cpsumon
