--- cp210x/Makefile
@@ -1,7 +1,0 @@
-obj-m = cp210x.o
-KVERSION = $(shell uname -r)
-all:
-	make -C /lib/modules/$(KVERSION)/build M=$(PWD) modules
-clean:
-	make -C /lib/modules/$(KVERSION)/build M=$(PWD) clean
-
