+++ include/cpsumoncli.h
@@ -1,0 +1,1 @@
+int main (int argc, char * argv[]);
