+++ libcpsumon/include/cpsumon.h
@@ -1,0 +1,122 @@
+/*
+Corsair AXi Series PSU Monitor
+Copyright (C) 2014 Andras Kovacs - andras@sth.sze.hu
+
+This program is free software; you can redistribute it and/or
+modify it under the terms of the GNU General Public License
+as published by the Free Software Foundation; either version 2
+of the License, or (at your option) any later version.
+
+This program is distributed in the hope that it will be useful,
+but WITHOUT ANY WARRANTY; without even the implied warranty of
+MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+GNU General Public License for more details.
+
+You should have received a copy of the GNU General Public License
+along with this program; if not, write to the Free Software
+Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
+
+*/
+#include <stdio.h>
+#include <string.h>
+#include <inttypes.h>
+#include <unistd.h>
+#include <stdlib.h>
+#include <errno.h>
+#include <fcntl.h>
+#include <termios.h>
+#include <math.h>
+
+#define TYPE_AX760  0
+#define TYPE_AX860  1
+#define TYPE_AX1200 2
+#define TYPE_AX1500 3
+
+#define true  1
+#define false 0
+
+typedef struct rail_12v_elem_t {
+    float voltage;
+    float current;
+    float power;
+    unsigned char ocp_enabled;
+    float ocp_limit;
+} rail_12v_elem_t;
+
+typedef struct rail_misc_elem_t {
+    float voltage;
+    float current;
+    float power;
+} rail_misc_elem_t;
+
+typedef struct rail_12v_t {
+    rail_12v_elem_t pcie[10];
+    rail_12v_elem_t atx;
+    rail_12v_elem_t peripheral;
+} rail_12v_t;
+
+
+typedef struct rail_misc_t {
+    rail_misc_elem_t rail_5v;
+    rail_misc_elem_t rail_3_3v;
+} rail_misc_t;
+
+typedef struct psu_main_power_t {
+    float voltage;
+    float current;
+    float inputpower;
+    float outputpower;
+    char cabletype;
+    float efficiency;
+} psu_main_power_t;
+
+static int _psu_type;
+static rail_12v_t _rail12v;
+static rail_misc_t _railmisc;
+static psu_main_power_t _psumain;
+
+
+unsigned char encode_table[16]  =
+			 {0x55, 0x56, 0x59, 0x5a, 0x65, 0x66, 0x69, 0x6a, 0x95, 0x96, 0x99, 0x9a, 0xa5, 0xa6, 0xa9, 0xaa};
+unsigned char decode_table[256] =
+			 {0x30, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+			  0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+			  0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+			  0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+			  0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x10,
+			  0x20, 0x21, 0x00, 0x12, 0x22, 0x23, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x14, 0x24,
+			  0x25, 0x00, 0x16, 0x26, 0x27, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+			  0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+			  0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x18, 0x28, 0x29, 0x00, 0x1a,
+			  0x2a, 0x2b, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x1c, 0x2c, 0x2d, 0x00, 0x1e, 0x2e,
+			  0x2f, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+			  0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+			  0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+			  0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+			  0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+			  0x00
+			};
+
+int xread(int f, void * b, int s, int timeout);
+int xwrite(int fd, void * buffer, unsigned int len);
+void dump(unsigned char *buf, int dlen);
+unsigned char * decode_answer(unsigned char *data, int size, int * nsize);
+unsigned char * encode_answer(unsigned char command, unsigned char *data, int size, int * nsize);
+float convert_byte_float(unsigned char * data);
+void convert_float_byte(float val, int exp, unsigned char *data);
+unsigned char * data_read_dongle(int fd, int size, int * command);
+int data_write_dongle(int fd, unsigned char * datain, int size);
+unsigned char * read_dongle_name(int fd);
+int read_dongle_version(int fd, float *f);
+unsigned char * read_data_psu(int fd, int reg, int len);
+unsigned char * write_data_psu(int fd, int reg, char * data, int len);
+int set_page(int fd, int main, int page);
+int set_12v_page(int fd, int page);
+int set_main_page(int fd, int page);
+int read_psu_main_power(int fd);
+int read_psu_rail12v(int fd);
+int read_psu_railmisc(int fd);
+int read_psu_fan_speed(int fd, float * f);
+int read_psu_temp(int fd, float * f);
+char * dump_psu_type(int type);
+int setup_dongle(int fd);
